improve code quality by reducing number of params
